(#20) feature(app): add asset folder and move svg icons files there
